refactor test helpers to be available for signifypy, all tests successful
